pyproject.toml
[project]
name = "demo2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
]
